+++ kinto/core/logs.py
@@ -1,12 +1,6 @@
-import os
+import logging
 
 import colorama
-import structlog
-
-from kinto.core import utils
-
-
-logger = structlog.get_logger()
 
 
 def log_context(request, **kwargs):
@@ -26,23 +20,22 @@ def decode_value(value):
         return bytes(value).decode('utf-8')
 
 
-class ClassicLogRenderer:
-    """Classic log output for structlog.
+class ColorFormatter(logging.Formatter):
+    EXCLUDED_LOGRECORD_ATTRS = set((
+        'args', 'asctime', 'created', 'exc_info', 'exc_text', 'filename',
+        'funcName', 'levelname', 'levelno', 'lineno', 'module', 'msecs',
+        'message', 'msg', 'name', 'pathname', 'process', 'processName',
+        'relativeCreated', 'stack_info', 'thread', 'threadName'
+    ))
 
-    ::
-
-        "GET   /v1/articles?_sort=title" 200 (3 ms) request.summary uid=234;
-
-    """
-    def __init__(self, settings):
-        pass
-
-    def __call__(self, logger, name, event_dict):
+    def format(self, record):
         RESET_ALL = colorama.Style.RESET_ALL
         BRIGHT = colorama.Style.BRIGHT
         CYAN = colorama.Fore.CYAN
         MAGENTA = colorama.Fore.MAGENTA
         YELLOW = colorama.Fore.YELLOW
+
+        event_dict = {**record.__dict__}
 
         if 'path' in event_dict:
             pattern = (BRIGHT +
@@ -52,10 +45,10 @@ class ClassicLogRenderer:
                        RESET_ALL +
                        ' {event} {context}')
         else:
-            pattern = '{event} {context}'
+            pattern = BRIGHT + '{event}' + RESET_ALL + ' {context}'
 
         output = {
-            'event': str(event_dict.pop('event', '?')).format(**event_dict)
+            'event': str(event_dict.pop('msg', '?')).format(**event_dict)
         }
         for field in ['method', 'path', 'code', 't']:
             output[field] = decode_value(event_dict.pop(field, '?'))
@@ -69,69 +62,9 @@ class ClassicLogRenderer:
             "=" +
             MAGENTA + decode_value(event_dict[key]) +
             RESET_ALL
-            for key in sorted(event_dict.keys())
+            for key in sorted([k for k in event_dict.keys()
+                               if k not in self.EXCLUDED_LOGRECORD_ATTRS])
         )
 
         log_msg = pattern.format_map(output)
         return log_msg
-
-
-class MozillaHekaRenderer:
-    """Build structured log entries as expected by Mozilla Services standard:
-
-    * https://mana.mozilla.org/wiki/display/CLOUDSERVICES/Logging+Standard
-    """
-
-    ENV_VERSION = '2.0'
-
-    def __init__(self, settings):
-        super().__init__()
-        self.appname = settings['project_name']
-        self.hostname = utils.read_env('HOSTNAME', os.uname()[1])
-        self.pid = os.getpid()
-
-    def __call__(self, logger, name, event_dict):
-        SYSLOG_LEVELS = {
-            'critical': 0,
-            'fatal': 0,
-            'exception': 2,
-            'error': 2,
-            'warning': 4,
-            'info': 6,
-            'debug': 7,
-        }
-        severity = SYSLOG_LEVELS[name]
-
-        MSEC_TO_NANOSEC = 1000000
-        timestamp = utils.msec_time() * MSEC_TO_NANOSEC
-
-        event = event_dict.pop('event', '')
-
-        defaults = {
-            'Timestamp': timestamp,
-            'Logger': self.appname,
-            'Type': event,
-            'Hostname': self.hostname,
-            'Severity': severity,
-            'Pid': self.pid,
-            'EnvVersion': self.ENV_VERSION,
-            'Fields': {}
-        }
-
-        for f, v in defaults.items():
-            event_dict.setdefault(f, v)
-
-        fields = [k for k in event_dict.keys() if k not in defaults]
-        for f in fields:
-            value = event_dict.pop(f)
-
-            # Heka relies on Protobuf, which doesn't support recursive objects.
-            if isinstance(value, dict):
-                value = utils.json.dumps(value)
-            elif isinstance(value, (list, tuple)):
-                if not all([isinstance(i, str) for i in value]):
-                    value = utils.json.dumps(value)
-
-            event_dict['Fields'][f] = value
-
-        return utils.json.dumps(event_dict)