+++ tests/core/listeners.py
@@ -1,7 +1,10 @@
 import json
+import logging
 
-from kinto.core import logger
 from kinto.core.listeners import ListenerBase
+
+
+logger = logging.getLogger(__name__)
 
 
 class Listener(ListenerBase):