+++ tests/core/test_authorization.py
@@ -187,9 +187,9 @@ class AuthorizationPolicyTest(unittest.TestCase):
         self.context.check_permission.return_value = False
         with mock.patch('kinto.core.authorization.logger') as mocked:
             self.authz.permits(self.context, self.principals, 'dynamic')
-        mocked.warn.assert_called_with('{userid} is not allowed to {perm} {uri}',
-                                       perm='read', uri='/articles/43/comments/2',
-                                       userid='portier:yourself')
+        mocked.warn.assert_called_with('Permission not granted.',
+                                       extra=dict(perm='read', uri='/articles/43/comments/2',
+                                                  userid='portier:yourself'))
 
     def test_permits_refers_to_context_to_check_permissions(self):
         self.context.check_permission.return_value = True